:triangular_flag_on_post: Enable MM encoding feature by default for a… (#1029)
…ll platform

<!-- markdownlint-disable -->

## Description

Enable separate encoder for power by default.

## Related Issues

<!-- Link related issues, e.g., `Fixes #` or `Relates to #456` -->

## Test Plan

<!-- Describe how you tested your changes. Include commands or steps to
reproduce. -->

## Checklist

- [ ] I have read the [contributing
guidelines](https://docs.vllm.ai/projects/spyre/en/latest/contributing)
- [ ] My code follows the project's code style (run `bash format.sh`)
- [ ] I have added tests for my changes (if applicable)
- [ ] I have updated the documentation (if applicable)
- [ ] My commits include a `Signed-off-by:` line (DCO compliance)

---------

+++ tests/utils/test_envs.py
@@ -48,7 +48,7 @@ def test_env_vars_override_for_invalid_config():
     "machine,expected",
     [
         ("s390x", torch.float32),
-        ("ppc64le", torch.bfloat16),
+        ("ppc64le", torch.float32),
         ("x86_64", torch.float16),
         ("aarch64", torch.float16),
     ],
